Replace unicode micro symbol (problematic when calling over ssh) and other minor tidy-ups

+++ src/cphTrace.c
@@ -1,6 +1,6 @@
-/*<copyright notice="lm-source" pids="" years="2007,2017">*/
+/*<copyright notice="lm-source" pids="" years="2007,2021">*/
 /*******************************************************************************
- * Copyright (c) 2007,2017 IBM Corp.
+ * Copyright (c) 2007,2021 IBM Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -391,7 +391,7 @@ int cphTraceListThreads(CPH_TRACE *pTrace) {
             CPH_ARRAYLISTITEM *pItem = cphListIteratorNext(pIterator);
             CPH_TRACETHREAD *pTraceThread = pItem->item;
 
-            printf("   pItem: %p, pTraceThread: %p, pTraceThread->threadId: %d.\n", pItem, pTraceThread, pTraceThread->threadId);
+            printf("   pItem: %p, pTraceThread: %p, pTraceThread->threadId: %d.\n", pItem, pTraceThread, (int)pTraceThread->threadId);
 
         } while (cphListIteratorHasNext(pIterator));
         cphListIteratorFree(&pIterator);
